modified

+++ apps/episode/views.py
@@ -1,5 +1,7 @@
+from django.shortcuts import get_object_or_404
 from django_filters import rest_framework
 from rest_framework import viewsets, generics
+from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 from rest_framework import filters
 from .serializers import (
@@ -96,11 +98,17 @@ class EpisodeLikeAPIView(generics.GenericAPIView):
 
     def post(self, request, *args, **kwargs):
         episode_id = self.kwargs['episode_id']
+        print(request.user)
         user_id = self.request.user.id
+        if not user_id:
+            raise ValidationError("You are not registered yet")
+        print(episode_id, user_id)
+        episode = get_object_or_404(Episode, pk=episode_id)
         has_like = EpisodeLike.objects.filter(episode_id=episode_id, author_id=user_id)
         if has_like:
             has_like.delete()
             return Response({'success': True, 'message': 'Episode like remove'})
         else:
-            EpisodeLike.objects.create(episode_id=episode_id, author_id=user_id)
+            EpisodeLike.objects.create(episode_id=episode.id, author_id=user_id)
             return Response({'success': True, 'message': 'Episode like add'})
+
